% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F0gin.R
\name{F0gin}
\alias{F0gin}
\title{Internal function
This is an internal function that adds two numbers.}
\usage{
F0gin(a, m, t, sign = TRUE, method = "Fortran")
}
\arguments{
\item{a}{a degrees-of-freedom parameter}

\item{m}{A single numeric value.}

\item{t}{parameter}

\item{sign}{default TRUE}

\item{method}{two ways to calculate pbdv}
}
\value{
F0gin
}
\description{
Internal function
This is an internal function that adds two numbers.
}
