% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mshift.R
\name{mshift}
\alias{mshift}
\title{Internal function}
\usage{
mshift(z, mode, a, m, shift = TRUE)
}
\arguments{
\item{z}{quantile}

\item{mode}{?}

\item{a}{a degrees-of-freedom parameter}

\item{m}{similar to location parameter, controls asymmetry of the distribution.}

\item{shift}{?}
}
\value{
mshift
}
\description{
This is an internal function that adds two numbers.
}
