% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbdv_r.R
\name{pbdv_r}
\alias{pbdv_r}
\title{title}
\usage{
pbdv_r(v, x, method = "Fortran")
}
\arguments{
\item{v}{intermediate parameters}

\item{x}{intermediate parameters}

\item{method}{intermediate parameters}
}
\value{
pbdv_r(v,x)
}
\description{
title
}
